LPS-189701 Merge BeanPortletRegistrarImpl into BeanPortletRegistrar and provide static methods, as BeanPortletRegistrarImpl does not need to be a component, it does not depend on other components, nor is it used in components
