--- modules/apps/bean-portlet/bean-portlet-registration-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Bean Portlet Registration Implementation
-Bundle-SymbolicName: com.liferay.bean.portlet.registration.impl
-Bundle-Version: 2.0.18
