--- modules/apps/bean-portlet/bean-portlet-registration-impl/build.gradle
@@ -1,14 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.apache.geronimo.specs", name: "geronimo-atinject_1.0_spec", version: "1.1"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
-	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
-	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
-	compileOnly project(":apps:bean-portlet:bean-portlet-api")
-	compileOnly project(":apps:bean-portlet:bean-portlet-extension-api")
-	compileOnly project(":apps:bean-portlet:bean-portlet-registration-api")
-	compileOnly project(":core:petra:petra-lang")
-	compileOnly project(":core:petra:petra-string")
-}
