+++ modules/modules.properties
@@ -2414,7 +2414,6 @@ project.name[com.liferay.bean.portlet.api]=:apps:bean-portlet:bean-portlet-api
 project.name[com.liferay.bean.portlet.cdi.extension]=:apps:bean-portlet:bean-portlet-cdi-extension
 project.name[com.liferay.bean.portlet.extension.api]=:apps:bean-portlet:bean-portlet-extension-api
 project.name[com.liferay.bean.portlet.registration.api]=:apps:bean-portlet:bean-portlet-registration-api
-project.name[com.liferay.bean.portlet.registration.impl]=:apps:bean-portlet:bean-portlet-registration-impl
 project.name[com.liferay.bean.portlet.spring.extension]=:apps:bean-portlet:bean-portlet-spring-extension
 project.name[com.liferay.blogs.api]=:apps:blogs:blogs-api
 project.name[com.liferay.blogs.demo.data.creator.api]=:apps:blogs:blogs-demo-data-creator-api
